modl/requirements.txt
absl-py==2.1.0
appnope==0.1.4
asttokens==2.4.1
clarabel==0.9.0
comm==0.2.2
contourpy==1.3.1
cvxpy==1.6.0
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
grpcio==1.68.0
h5py==3.12.1
imageio==2.36.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
osqp==0.6.7.post3
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
qdldl==0.1.7.post4
scikit-image==0.24.0
scipy==1.14.1
scs==3.2.7
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tifffile==2024.9.20
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.1.3
